Show more output when running docker-compose
Currently we don't show the docker-compose progress (due to no TTY) which means there's no progress on the initial images pull which isn't not at all helpful, as it can really take some time when doing the first start on a machine.

Also, I added the same to the proxy command, as that has the same issue.