Don't send empty cookie on every request
When not using `PHP_XDEBUG_ENABLED`, we currently send a blank cookie on _every_ request. This is wasteful, but also means you won't get Batcache locally because of cookies being set, and can snag other problems too.

I don't really see why we even need to set this at all. In this context.