Network PHP container to allow access to service hostnames
Now PHP uses sockets instead of port 9000 for the Nginx -> PHP connection, we can put PHP on the main proxy network. This means it's then easy to link in the hostnames of our services to the traefik HTTP server.

This means we can make requests from PHP to https://project.altis.dev/ as well as https://pinpoint-project.altis.dev/ etc.