Use Minio for S3 storage
This update switches from fakes3 to minio plus the minio client (mc). This allows us to properly support virtualhost style routing as per AWS where the subdomain selects the bucket.

By using the `mc` docker image we can remove some not too robust code that was designed to ensure the bucket was created by uploading a file and then removing it.

Additionally we can use `mc` to run a synchronisation service. Currently I have this running to sync from s3 to `content/uploads`. During startup eg. `composer serve` any existing files in `content/uploads` are synced to s3. Unfortunately getting this to work well in a bidirectional way is almost impossible. So we may want to think about our preferred direction to do synchronisation by default and perhaps even add a command to run the sync from the host to s3.