Add support for Signed Urls in Tachyon
Currently we have  mix between using path-type and subdomain-style S3 bucket urls in Local Server. Because Tachyon, and the WordPress application use different URLs patterns, it means the Signed URLs generated between them are not compatible (the url path is part of the signature, so the path has to be the same in both).

We've sinse added support for loop-back requests to the S3 container on the proxy network, so we can now use the "s3-$project.altis.dev" hostname from the Tahcyon container, matching the hostname we use from the WordPress applicatio container. It also means we don't need to change the behaviour of the S3 SDK to special-case local development.