Generate docker-compose.yml file in PHP
We are reaching the limitations of what we can do with a static docker-compose file. There's no way to optionally disable services for example, or provide configuration for the containers from the `composer.json` or CLI flags. This moves the definition of the docker-compose file into PHP, and will dump it out to a docker-compose.yml file for use in `docker-compose`.