Support for ES packages
Overrides the stored package IDs locally.

Changes the S3 sync to overwrite to enable updates from S3 when package files are updated.

Mounts the `content/uploads/search` directory into the ES config directory so that files can be referenced in a consistent way with how they are referenced on AWS