Improve file sharing perforamnce with single volume
Note this is only valid for Mac and Linux currently.

It creates a top level volume for docker-compose which is shared by the services that need to access. In theory this means we are syncing once rather than the same stuff 3 times. In local development I have seen built files fail to sync for long periods of time requiring restarting the PHP container. It is painful to say the least.