Add a create-alias command to local server
This adds a `create-alias`` command to local server that will generate a
wp-cli.local.yml file in the root of the project. This file will contain an
alias for the php docker container that can be used with wp-cli commands.

The command is conservative in that it will not overwrite an existing wp-cli.local.yml file.

This also add an alias to the `cli` command so that you can use `composer server wp` too.