Always install the root certs
This changes fixes ticket https://github.com/humanmade/altis-local-server/issues/566

Before this change: If `mkcert` was already installed the code did not also install the local CA in the system trust store. (using `mkcert -install`).

After this change: If `mkcert` is installed globally or locally, it always goes on to run `mkcert -install` and install the local CA in the system trust store.